chore: improve Fragment exports
